feat(pydantic): organize code
